Fetch the latest repository-list.txt directly from the github
This will make things easier when we merge/remove some of the repos.
So far we had to manually fix the lists in the plugin, then upgrade
the plugin and restart jenkins, which is quite cumbersome for such
a small change.

One remaining point is to fix this properly for PRs which change the
repository-list.txt as well, so that all the downstream PRs actually
use the updated list as well (and not the "mail" one which is outdated
for those PRs).