[#6] read the branch mapping configuration from external file
Before these changes it was necessary to update the plugin source code
everytime new UF/Dash/KIE was created. That also led to a new upload of
the plugin and Jenkins restart. That approach was disruptive and error
prone.

This commit eases the update process as it does not hard code any branch
mapping in the code, but rather uses an external file which contains
those mappings. From now on, when creating a new branch it should be
only necessary to update that external file.